Embed a X509 certificate when signing documents.
The module was embedding a PEM-formatted public key as the value of
the <X509Certificate> element, rather than the certificate. The
public key was derived from the private key, so this change adds a
new "cert" option to allow the corresponding certificate to be provided.

This change permits the TODO in t/006_signing.t to be removed as the
test now passes in its entirety.