Add the canonicalization transform we do when signing.
We canonicalize before signing, so we need to add that transform
to the set in our SignedInfo. This causes the verify operation to
also canonicalize.

Update the signing test to show that even if the document to be
signed is not already in canonical form we can successfully round-
trip it.