Update testsuite
In an older branch of mine I had some updates on the testsuite to make
it more useful. It mainly tries to move away from `ok($thing->func)` or
`ok('foo' eq 'bar')` syntax and actually testing the outcome of
whatever was called. In addition there are some testsuite helpers.

There are a couple of things that are different between .17 and .20 and
up. The tests in t/07-logout-request.t proved to be valuable, in my
previous branch everything worked. Some issues occured:

4e3ff81d4f2b963a855a5c24c96c87e2d4577c5b - this commit wipes the
NameIDPolicy in the XML from Net::SAML2::Protocol::AuthnRequest
(nameid_format attribute is now simply ignored). This is a breaking
change for consumers of .17

7f459c303c250c532d5b47f8e137bdd88c4fc2a7 - this commit reintroduces it,
but under a different name: nameidpolicy_format. Now older code will
need to deal with this change. I therefore introduced an `around
BUILDARGS` which changes `nameid_format` to `nameidpolicy_format`.
Although I think we should pick *one* and stick to it. Other code uses
`nameid_format`: Net::SAML2::Protocol::LogoutRequest, Net::SAML2::SP and
Net::SAML2.

AllowCreate was completely ignored, this used to be true by default, but
now has an attribute with a predicate. If you set it, we'll use it,
otherwise it is ignored. This is a (potential) breaking change for
anyone who used 0.17 for the last couple of years.

There are no other code changes in this PR, other than in the testsuite.
Hopefully this will make it easier to refactor code in the future.

Signed-off-by: Wesley Schwengle <wesley@opndev.io>