Initial full client test
This test follows the flow of the TUTORIAL.MD.  The comments below
follows the major headings of that document to implement a full
SP client authentication process.

The only part that is missing is the user, redirect, IdP
web server flow.  That is implemented via a call to
Test::Net::SAML2->net_saml2_binding_redirect_request($url, $idp).

The function receives the url that would be used to access the IdP and
responds with the SAMLResponse that the web browser would receive back
from the IdP.  Essentially implementing the login process at the IdP.