Remove setting default format when not present in metadata
The SAML spec says:
<NameIDFormat> [Zero or More]
Zero or more elements of type anyURI that enumerate the name identifier formats supported by
this system entity acting in this role. See Section 8.3 of [SAMLCore] for some possible values for
this element.

In the function format we claim that when nothing is set, you get undef. The
current code always sends a name-id-format to the caller even if the metadata
did not provide any data.

Signed-off-by: Wesley Schwengle <wesley@opndev.io>