Add the index in the assertion_consumer_service
This way you can later ask the SP which one it is with the
`get_default_assertion_service` call. This is handy for those who configure the
SP and later on need to know the index of the assertion consumer service. Some
IdP's require this in the AuthnRequest and this way you can easily query the
SP.

Signed-off-by: Wesley Schwengle <wesley@opndev.io>