Set version in main module for downstream developers
By setting the version in lib/Net/SAML2.pm downstream developers can use a git
repo to install Net::SAML2 and still keep their dependencies correct. If a
module requires Net::SAML2 v0.56 and the version in git is not set, it cannot
validate the module version and it breaks the versioning check of for example
cpanminus. By setting the version in the main module we allow downstream users
to test/play with development versions of Net::SAML2.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>