Implement verify_xml() call which only verifies the XML
BREAKING CHANGE: handle_reponse returns different data

The verify_xml() call ONLY verifies the XML and does nothing with the name of
the subject anymore. The implementation was a bit odd. When a CA-cert was given
the subject was extracted from the cert and returned, but when a CA-cert was
not given the reponse was a true value.

handle_response in both Net::SAML2::Binding::POST and SOAP only return the
response XML and not an array with the subject and the response XML.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>