Make attributes in NameID optional
The Format attribute in NameID will only be set if there is a nameid format
configured. The NameQualifier and SPNameQualifier are only set when this is
configured by setting `include_name_qualifier`. This is because the SAML spec
says:

> The NameQualifier and SPNameQualifier attributes SHOULD be omitted unless the
> element or format explicitly defines their use and semantics.

The SAML spec also says that if the NameID Format is
`urn:oasis:names:tc:SAML:2.0:nameidformat:persistent` tese two options must be
set. We support this. If there is an affiliation_group_id than we use that
instead of the destination ID because of the same spec.

> In any case, the <saml:NameID> content in the request and its associated
> SPProvidedID attribute MUST contain the most recent name identifier
> information established between the providers for the principal.
>
> In the case of an identifier with a Format of
> urn:oasis:names:tc:SAML:2.0:nameidformat:persistent, the NameQualifier
> attribute MUST contain the unique identifier of the identity provider that
> created the identifier. If the identifier was established between the
> identity provider and an affiliation group of which the service provider is a
> member, then the SPNameQualifier attribute MUST contain the unique identifier
> of the affiliation group. Otherwise, it MUST contain the unique identifier of
> the service provider. These attributes MAY be omitted if they would otherwise
> match the value of the containing protocol message's <Issuer> element, but
> this is NOT RECOMMENDED due to the opportunity for confusion.

For more information see:
http://docs.oasis-open.org/security/saml/v2.0/saml-core-2.0-os.pdf

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>