Always return a default assertion service
http://docs.oasis-open.org/security/saml/v2.0/saml-metadata-2.0-os.pdf
2.2.3 Complex Type IndexedEndpointType

isDefault [Optional]

An optional boolean attribute used to designate the default endpoint among an
indexed set. If omitted, the value is assumed to be false.

In any such sequence of like endpoints based on this type, the default endpoint
is the first such endpoint with the isDefault attribute set to true. If no such
endpoints exist, the default endpoint is the first such endpoint without the
isDefault attribute set to false. If no such endpoints exist, the default
endpoint is the first element in the sequence

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>