Split Github Actions workflow
We supported perl 5.12 - perl 5.38, which are about 14 versions of Perl. The CI
ran tests for all of them and if one of them failed, they all failed. This is a
weebit anoying. Now we split the CI in three sections:

1. We test the last 3 versions of perl
2. We test every version which does not require Debian buster
3. We test every version which requires Debian buster

This allows us to first test the latest versions of Perl and will later fail if
older versions fail.

Additionally, bump Perl version to 5.14 as IO::Socket::IP bumped its version
requirement as well[^1][^2].

[^1]: https://rt.cpan.org/Public/Bug/Display.html?id=149138
[^2]: https://metacpan.org/dist/IO-Socket-IP/changes

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>