Rename id to issuer for Net::SAML2::SP
We've had generate_sp_desciptor_id to override the ID, but this could
very well become a parameter for the constructor. The current ID
attribute was actually used as the issuer. Rename ID to issuer and and
reintroduce ID. We add some deprecation warnings to users so they are
aware of the change.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>