Fix signature location bug
In the XSD of SAML metadata the signature should be present directly
after the md:EntityDescriptor tag as a child node. We placed it at the
bottom of the field violating the XSD of SAML metadata.

We should potentially fix it in XML::Sig but that is for another day.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>