Correctly test NameID attributes
In 6c63ed1 we introduced some logic to make NameID attributes optional.
On the checks is that urn:oasis:names:tc:SAML:2.0:nameidformat:persistent is
special. There are two problems with this, one:
urn:oasis:names:tc:SAML:2.0:nameidformat:persistent should be
urn:oasis:names:tc:SAML:2.0:nameid-format:persistent. And two: the test in
t/07-logout-request.t should test that SPNameQualifier and NameQualifier are
actually set and they are now not set (for the test that is). By fixing
the typo, we need to change the tests to test the correct thing.

Closes: #213

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>